Move the shared code in fs/decomposed to utils
